examples/demoapp/requirements.txt
networkx
gradio
pyjwt